Add basic library framework and test executable
The library provides functionality to create a context and add
windows to it. The context then processes a frame and outputs a
number of vertices to draw the user interface. This functionality
is exposed via a single header file for convenience (ModernUI.h).

The test executable needs to be enabled with the CMake option
MODERNUI_BUILD_TEST and requires OpenGL 3.3 support. It shows how
an application can use the library and render the user interface
to the screen.