Fix compilation on Windows
The external dependencies glad and glfw needed to be updated for
Windows. A very large array on the stack was also moved to the
heap to avoid an overflow on Windows. The PIMPL pattern usage was
improved by way of unique pointers for automatic memory
management. All default-level warnings were also removed.